requirements.txt
Pillow
text-to-image
disnake
discordspy
discord.py
qrcode
topggpy
requests
numba
